refactor: move Footer under components folder and add link to rocket-coding website
